fix set_submodule and add replace_or_create_new_leaf_module
before fixing set_submodule, it used to create leaf modules when the target was not a dot-delimited string. after the fix it will not create a new attribute if target is a non-dot-delimited string. If you want to create leaf nodes of nn.Module parent nodes, you can use replace_or_create_new_leaf_module